requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Frontend
streamlit>=1.28.0

# Data Processing
datasets>=2.14.0
sentencepiece>=0.1.99

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.1.0
pyyaml>=6.0.1
requests>=2.31.0
huggingface-hub>=0.19.0

